Library: Better error handling
