+++ fpjs-pro/src/main/java/com/fingerprintjs/android/fpjs_pro/device_id_providers/AndroidIdProvider.kt
@@ -7,7 +7,7 @@ import android.provider.Settings
 import com.fingerprintjs.android.fpjs_pro.tools.executeSafe
 
 
-class AndroidIdProvider(
+internal class AndroidIdProvider(
     private val contentResolver: ContentResolver
 ) {
     @SuppressLint("HardwareIds")
