+++ fpjs-pro/src/main/java/com/fingerprintjs/android/fpjs_pro/tools/ExceptionSafeExecutor.kt
@@ -4,7 +4,7 @@ package com.fingerprintjs.android.fpjs_pro.tools
 import java.lang.Exception
 
 
-fun <T> executeSafe(code: () -> T, defaultValue: T): T {
+internal fun <T> executeSafe(code: () -> T, defaultValue: T): T {
     return try {
         code()
     } catch (exception: Exception) {
